task user crud relation

+++ src/tasks/tasks.service.ts
@@ -5,6 +5,8 @@ import {TaskRepository} from "./task.repository";
 import {InjectRepository} from "@nestjs/typeorm";
 import {Task} from "./task.entity";
 import {TaskStatus} from "./task-status.enum";
+import {GetUser} from "../auth/get-user.decorator";
+import {User} from "../auth/user.entity";
 
 @Injectable()
 export class TasksService {
@@ -17,12 +19,18 @@ export class TasksService {
 
     tasks = [];
 
-    async getTasks(filterDto: GetTasksFilterDto): Promise<Task[]> {
-        return this.taskRepository.getTasks(filterDto);
+    async getTasks(
+        filterDto: GetTasksFilterDto,
+        @GetUser() user: User
+    ): Promise<Task[]> {
+        return this.taskRepository.getTasks(filterDto, user);
     }
 
-    async getTaskById(id: number): Promise<Task> {
-        const task = await this.taskRepository.findOne(id);
+    async getTaskById(
+        id: number,
+        @GetUser() user: User
+    ): Promise<Task> {
+        const task = await this.taskRepository.findOne({where: { id, userId: user.id }});
 
         if(!task){
             throw new NotFoundException(`Task with ID: ${id} not found`);
@@ -31,20 +39,26 @@ export class TasksService {
         return task;
     }
 
-    async createTask(createTaskDto: CreateTaskDto): Promise<Task>{
-        return this.taskRepository.createTask(createTaskDto);
+    async createTask(
+        createTaskDto: CreateTaskDto,
+        @GetUser() user: User
+    ): Promise<Task>{
+        return this.taskRepository.createTask(createTaskDto, user);
     }
 
 
-    async updateTaskStatus(id: number, status: TaskStatus): Promise<Task> {
-        const task = await this.getTaskById(id);
+    async updateTaskStatus(id: number, status: TaskStatus, user: User): Promise<Task> {
+        const task = await this.getTaskById(id, user);
         task.status = status;
         await task.save();
         return task;
     }
 
-    async deleteTask(id: number): Promise<void>{
-        const result = await this.taskRepository.delete(id);
+    async deleteTask(
+        id: number,
+        @GetUser() user: User
+    ): Promise<void>{
+        const result = await this.taskRepository.delete({id, userId: user.id});
 
         if(result.affected === 0){
             throw new NotFoundException(`Task with ID: ${id} not found`);
